Function strutil::is_anagram

pub fn is_anagram(subject: &str, string: &str) -> bool

Determines if the provided string is an anagram of the supplied subject.

Rules

strings are of same length

an exact character mapping between strings, lowercased and special characters counted.

Examples

extern crate strutil;

let subject = "Eleven plus two";
let anagram = "Twelve plus one";

assert_eq!(true, strutil::is_anagram(&subject, &anagram))